+++ chimera/render/PVS.hpp
@@ -1,6 +1,7 @@
 #ifndef __CHIMERA_PVS_HPP
 #define __CHIMERA_PVS_HPP
 
+#include "Plane.hpp"
 #include "Triangle.hpp"
 #include <vector>
 
@@ -11,7 +12,7 @@ struct Node {
     unsigned long Plane;
     unsigned long Front;
     signed long Back;
-    //BOUNDINGBOX BoundingBox;
+    // BOUNDINGBOX BoundingBox;
 };
 
 class Tris {
@@ -33,7 +34,7 @@ class PVS {
     PVS() {}
     unsigned int selectBestSplitter(std::vector<Tris>& _poliyList);
 
-    std::vector<PlanePoint*> planes;
+    std::vector<Plane*> planes;
 };
 
 } // namespace Chimera
