+++ chimera/src/render/Triangle.cpp
@@ -40,84 +40,4 @@ void Triangle::generateNormal() {
     for (int i = 0; i < 3; i++)
         this->vertex[i].normal = normal;
 }
-
-void PlanePoint::set(const glm::vec3& _position, const glm::vec3& _normal) {
-    point = _position; //_point->vertex[0].position;
-    normal = _normal;  //_point->normal();
-}
-
-SIDE PlanePoint::classifyPoint(glm::vec3* point) {
-    // ref: http://www.cs.utah.edu/~jsnider/SeniorProj/BSP/default.htm
-    float clipTest;
-    glm::vec3 dir = this->point - (*point);
-    clipTest = glm::dot(dir, this->normal);
-
-    if (clipTest < -EPSILON)
-        return SIDE::CP_FRONT;
-
-    if (clipTest > EPSILON)
-        return SIDE::CP_BACK;
-
-    return SIDE::CP_ONPLANE;
-}
-
-SIDE PlanePoint::classifyPoly(const glm::vec3& pA, const glm::vec3& pB, const glm::vec3& pC, glm::vec3* clipTest) {
-    // ref: http://www.cs.utah.edu/~jsnider/SeniorProj/BSP/default.htm
-    unsigned short infront = 0;
-    unsigned short behind = 0;
-    unsigned short onPlane = 0;
-    float result[3];
-
-    result[0] = glm::dot((this->point - pA), this->normal); // Clip Test poin A
-    result[1] = glm::dot((this->point - pB), this->normal); // Clip Test poin B
-    result[2] = glm::dot((this->point - pC), this->normal); // Clip Test poin C
-
-    for (unsigned short a = 0; a < 3; a++) {
-        if (result[a] > EPSILON) {
-            behind++;
-        } else if (result[a] < -EPSILON) {
-            infront++;
-        } else {
-            onPlane++;
-            infront++;
-            behind++;
-        }
-    }
-
-    clipTest->x = result[0];
-    clipTest->y = result[1];
-    clipTest->z = result[2];
-
-    if (onPlane == 3)
-        return SIDE::CP_ONPLANE;
-
-    if (behind == 3)
-        return SIDE::CP_BACK;
-
-    if (infront == 3)
-        return SIDE::CP_FRONT;
-
-    return SIDE::CP_SPANNING;
-}
-
-bool PlanePoint::intersect(glm::vec3* linestart, glm::vec3* lineend, glm::vec3* intersection, float* percentage) {
-
-    glm::vec3 direction = (*lineend) - (*linestart);
-    float linelength = glm::dot(direction, this->normal);
-    if (fabsf(linelength) < 0.0001)
-        return false;
-
-    glm::vec3 L1 = this->point - (*linestart);
-
-    float dist_from_plane = glm::dot(L1, this->normal);
-    *percentage = dist_from_plane / linelength;
-
-    if (*percentage < 0.0f)
-        return false;
-    else if (*percentage > 1.0f)
-        return false;
-
-    *intersection = (*linestart) + (direction * (*percentage));
-    return true;
-}
 } // namespace Chimera
